+++ enclave/app/src/salary.rs
@@ -2,7 +2,7 @@ use indexmap::IndexMap;
 use uuid::Uuid;
 
 pub struct Salary {
-    pub salary_map: IndexMap<String, u16>
+    pub salary_map: IndexMap<String, u32>
 }
 
 impl Salary {
@@ -13,7 +13,7 @@ impl Salary {
         }
     }
 
-    pub fn add(&mut self, salary: u16) -> String {
+    pub fn add(&mut self, salary: u32) -> String {
         let id = Uuid::new_v4().to_string();
         self.salary_map.insert(id.clone(), salary);
         self.sort_salary_map();
@@ -25,9 +25,9 @@ impl Salary {
         self.salary_map.sort_by(|_a_key, a_value, _b_key, b_value| b_value.cmp(a_value));
     }
 
-    pub fn get_position(&self, id: String) -> Option<u16> {
+    pub fn get_position(&self, id: String) -> Option<u32> {
         match self.salary_map.get_index_of(&id) {
-            Some(index) => Some((index as u16) + 1),
+            Some(index) => Some((index as u32) + 1),
             None => None
         }
     }
